\name{recoup-defunct}
\alias{recoup-defunct}
\title{Defunct functions in package \sQuote{recoup}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{recoup} only, and will be defunct at the next release.
}

\details{
  The following functions are defunct and will be made defunct; use
  the replacement indicated below:
  \itemize{
    \item{coverageRnaRef: \code{\link{coverageRef}}}
  }
}
