\docType{data}
\name{test.design}
\alias{test.design}
\title{Design file for recoup test data}
\format{a \code{data.frame} whose format is accepted by recoup.}
\source{
    Personal communication with the Talianids lab at BSRC 
    'Alexander Fleming'. Unpublished data.
}
\description{
    This data frame contains 100 genes categorized 
    according to expression and strand.
}
\author{
    Panagiotis Moulos
}
\keyword{datasets}
