\docType{data}
\name{test.genome}
\alias{test.genome}
\title{A small gene set (100 genes) for recoup testing}
\format{a \code{data.frame} whose format is accepted by recoup.}
\source{
    Ensembl
}
\description{
    This gene set contains 100 genes and their coordinates 
    from mouse mm9 chromosome 12.
}
\author{
    Panagiotis Moulos
}
\keyword{datasets}

