\docType{data}
\name{test.input}
\alias{test.input}
\title{Input test ChIP-Seq data for recover}
\format{a \code{list} whose format is accepted by recover.}
\source{
    Personal communication with the Talianids lab at BSRC 
    'Alexander Fleming'. Unpublished data.
}
\description{
    This data set contains 10000 reads from H4K20me1 
    ChIP-Seq data from WT adult mice and Set8 (Pr-Set7)
    KO mice. The tissue is liver.
}
\author{
    Panagiotis Moulos
}
\keyword{datasets}

