\docType{data}
\name{test.exons}
\alias{test.exons}
\title{Genomic regions for recoup RNA-Seq plot testing}
\format{a \code{GRangesList}.}
\source{
    Ensembl
}
\description{
    This \code{GRangesList} contains the exons of 100 genes 
    for use with \code{recoup} RNA-Seq mode.
}
\author{
    Panagiotis Moulos
}
\keyword{datasets}
